test: wait for DevTools restart in redeployment ITs
Hoist ThemeSwitchLiveReloadIT's checkIfServerAvailable wait into
AbstractReloadIT so SessionValueIT and DevModeClassCacheIT also tolerate
a Spring Boot DevTools restart triggered by a neighbouring test.
